+++ components/common/data/src/PressureSensorData.hpp
@@ -18,7 +18,7 @@ public:
     float temperature = 0.0f;
 
     // Metadata
-    float timestamp = 0;
+    uint32_t timestamp = 0;
 
     json toJSON() const override {
         json j;
